Reworked the way functions are called. Added lots of tests.
So if a call argument is a function call, it has its names substituted
before the args are applied to the outer function, so it has a function
definition from the appropriate namespace, and the outer function then
receives a resolved function call, which can be used both: in an if
statement test expressions, in which case an attempt is made to call
the function and evaluate the returned expression; and in the function
body, when the passed function call is called as if had been defined
within the function and not passed as an argument to it.

Then had to fix call cache, so branching calls recombine effectively.

Also fixed multi-threading app, to handle interrupt signal.